Move interceptor logic into sepparate file
