+++ app/api/axios.js
@@ -1,7 +1,6 @@
 import Axios from "axios";
 import https from "https";
-import { parseJwt, getCookie } from "@/app/utils/cookieUtils.js";
-
+import { setupSessionInterceptors } from "./interceptors";
 //TODO: Get the base url from the env
 const BASE_URL =
   process.env.NEXT_PUBLIC_NODE_ENV === "production"
@@ -18,7 +17,7 @@ export const axios = Axios.create({
   headers: {
     "Content-Type": "application/json",
   },
-  withCredentials: true,
+  //withCredentials: true, TODO: Uncomment when we setup CORS
 });
 
 //TODO: Add axios private instance for the auth
@@ -35,7 +34,7 @@ export const axiosTour = Axios.create({
   headers: {
     "Content-Type": "multipart/form-data",
   },
-  withCredentials: true,
+  //withCredentials: true, TODO: Uncomment when we setup CORS
   httpsAgent: agent, // to be removed for production
 });
 
@@ -47,22 +46,7 @@ export const axiosAdmin = Axios.create({
   httpsAgent: agent, // to be removed for production
 });
 
-const attachTokenInterceptor = (instance) => {
-  instance.interceptors.request.use(
-    (config) => {
-      const session = getCookie("session");
-
-      if (session && session.accessToken) {
-        config.headers["Authorization"] = `Bearer ${session.accessToken}`;
-      }
-      return config;
-    },
-    (error) => {
-      return Promise.reject(error);
-    }
-  );
-};
-
-// Attach the interceptor to axiosTour and axiosAdmin
-attachTokenInterceptor(axiosTour);
-attachTokenInterceptor(axiosAdmin);
+// Attach the interceptor to axios, axiosTour and axiosAdmin
+setupSessionInterceptors(axios);
+setupSessionInterceptors(axiosTour);
+setupSessionInterceptors(axiosAdmin);
